Type promotions

+++ src/irvine/factor/util/FactorSequence.java
@@ -480,7 +480,7 @@ public final class FactorSequence {
     Z prod = Z.ONE;
     for (final Map.Entry<Z, Factor> f : mFactors.entrySet()) {
       final Z p = f.getKey();
-      prod = prod.multiply(p.pow((f.getValue().mExponent + 1) * degree).subtract(1)).divide(p.pow(degree).subtract(1));
+      prod = prod.multiply(p.pow((f.getValue().mExponent + 1) * (long) degree).subtract(1)).divide(p.pow(degree).subtract(1));
     }
     return prod;
   }
